Add `NewDigest`, `NewDigest64`, `NewDigest128` (#29)
The `New`, `New64` and `New128` functions return a `hash.Hash` interface
from the standard library, but this prevents Go's heap escape analysis from
being able to determine whether or not the arguments to the `Sum` functions
should escape to the heap or not, therefore they _always_ escape.

For example, when trying to avoid the heap with a stack array allocation:

```
hh, _ := highwayhash.New64(...)
var hb [highwayhash.Size64]byte
if !bytes.Equal(hh.Sum(hb[:0]), ...) { ... }
```

... the `hash.Hash` type for `hh` guarantees that `hb` will always escape
whereas a concrete return type allows the heap escape analysis to run
correctly and avoid the heap escape.

Signed-off-by: Neil Twigg <git@neilalexander.dev>